# Faster optimized frozen dictionary creation (4/n)