Instead of creating a dictionary of lists and a multi-dimensional array we rent a single dimension array, where every bucket has five slots.
The bucket starts at `(key.Length - minLength) * 5` index of the array.
 Each value is an index of the key from _keys array or just -1, which represents "null".
We avoid having two copies and re-ordering of keys and values collections.

Creation time is from x2 to x10 times faster, not more than 2x slower compared to Dictionary.
Indexing is 6-12% slower, but for large inputs it's still and order of magnitude faster than Dictionary.

|                                     Method |        Job | Count | ItemsPerBucket |             Mean | Ratio | Allocated | Alloc Ratio |
|------------------------------------------- |----------- |------ |--------------- |-----------------:|------:|----------:|------------:|
|                               ToDictionary |     #87688 |    10 |              1 |        102.26 ns |     - |     440 B |           - |
|               ToFrozenDictionary_Optimized |     #87876 |    10 |              1 |        218.24 ns |  0.25 |     488 B |        0.16 |
|               ToFrozenDictionary_Optimized |     #87688 |    10 |              1 |        889.59 ns |  1.00 |    3120 B |        1.00 |
|                                            |            |       |                |                  |       |           |             |
|                TryGetValue_True_Dictionary |     #87688 |    10 |              1 |         97.78 ns |     - |         - |          NA |
| TryGetValue_True_FrozenDictionaryOptimized |     #87876 |    10 |              1 |         24.48 ns |  1.13 |         - |          NA |
| TryGetValue_True_FrozenDictionaryOptimized |     #87688 |    10 |              1 |         21.69 ns |  1.00 |         - |          NA |
|                                            |            |       |                |                  |       |           |             |
|                               ToDictionary |     #87688 |    10 |              5 |        104.26 ns |     - |     440 B |           - |
|               ToFrozenDictionary_Optimized |     #87876 |    10 |              5 |        223.41 ns |  0.43 |     328 B |        0.33 |
|               ToFrozenDictionary_Optimized |     #87688 |    10 |              5 |        523.13 ns |  1.00 |    1000 B |        1.00 |
|                                            |            |       |                |                  |       |           |             |
|                TryGetValue_True_Dictionary |     #87688 |    10 |              5 |         98.22 ns |     - |         - |          NA |
| TryGetValue_True_FrozenDictionaryOptimized |     #87876 |    10 |              5 |         90.22 ns |  1.12 |         - |          NA |
| TryGetValue_True_FrozenDictionaryOptimized |     #87688 |    10 |              5 |         80.66 ns |  1.00 |         - |          NA |
|                                            |            |       |                |                  |       |           |             |
|                               ToDictionary |     #87688 |   100 |              1 |        766.64 ns |     - |    3128 B |           - |
|               ToFrozenDictionary_Optimized |     #87876 |   100 |              1 |        995.67 ns |  0.14 |    3728 B |        0.12 |
|               ToFrozenDictionary_Optimized |     #87688 |   100 |              1 |      7,179.71 ns |  1.00 |   30320 B |        1.00 |
|                                            |            |       |                |                  |       |           |             |
|                TryGetValue_True_Dictionary |     #87688 |   100 |              1 |      1,761.21 ns |     - |         - |          NA |
| TryGetValue_True_FrozenDictionaryOptimized |     #87876 |   100 |              1 |        227.97 ns |  1.13 |         - |          NA |
| TryGetValue_True_FrozenDictionaryOptimized |     #87688 |   100 |              1 |        201.72 ns |  1.00 |         - |          NA |
|                                            |            |       |                |                  |       |           |             |
|                               ToDictionary |     #87688 |   100 |              5 |        761.50 ns |     - |    3128 B |           - |
|               ToFrozenDictionary_Optimized |     #87876 |   100 |              5 |        996.27 ns |  0.25 |    2128 B |        0.24 |
|               ToFrozenDictionary_Optimized |     #87688 |   100 |              5 |      3,959.95 ns |  1.00 |    8768 B |        1.00 |
|                                            |            |       |                |                  |       |           |             |
|                TryGetValue_True_Dictionary |     #87688 |   100 |              5 |      1,042.10 ns |     - |         - |          NA |
| TryGetValue_True_FrozenDictionaryOptimized |     #87876 |   100 |              5 |        929.69 ns |  1.08 |         - |          NA |
| TryGetValue_True_FrozenDictionaryOptimized |     #87688 |   100 |              5 |        864.51 ns |  1.00 |         - |          NA |
|                                            |            |       |                |                  |       |           |             |
|                               ToDictionary |     #87688 |  1000 |              1 |      7,234.20 ns |     - |   31016 B |           - |
|               ToFrozenDictionary_Optimized |     #87876 |  1000 |              1 |      8,918.56 ns |  0.12 |   36128 B |        0.12 |
|               ToFrozenDictionary_Optimized |     #87688 |  1000 |              1 |     77,083.15 ns |  1.00 |  302344 B |        1.00 |
|                                            |            |       |                |                  |       |           |             |
|                TryGetValue_True_Dictionary |     #87688 |  1000 |              1 |    115,510.37 ns |     - |         - |          NA |
| TryGetValue_True_FrozenDictionaryOptimized |     #87876 |  1000 |              1 |      2,559.60 ns |  1.09 |         - |          NA |
| TryGetValue_True_FrozenDictionaryOptimized |     #87688 |  1000 |              1 |      2,341.85 ns |  1.00 |         - |          NA |
|                                            |            |       |                |                  |       |           |             |
|                               ToDictionary |     #87688 |  1000 |              5 |      7,541.47 ns |     - |   31016 B |           - |
|               ToFrozenDictionary_Optimized |     #87876 |  1000 |              5 |      9,438.56 ns |  0.24 |   20128 B |        0.23 |
|               ToFrozenDictionary_Optimized |     #87688 |  1000 |              5 |     39,774.73 ns |  1.00 |   88040 B |        1.00 |
|                                            |            |       |                |                  |       |           |             |
|                TryGetValue_True_Dictionary |     #87688 |  1000 |              5 |     31,294.80 ns |     - |         - |          NA |
| TryGetValue_True_FrozenDictionaryOptimized |     #87876 |  1000 |              5 |      9,168.88 ns |  1.12 |         - |          NA |
| TryGetValue_True_FrozenDictionaryOptimized |     #87688 |  1000 |              5 |      8,156.66 ns |  1.00 |         - |          NA |
|                                            |            |       |                |                  |       |           |             |
|                               ToDictionary |     #87688 | 10000 |              1 |    112,776.03 ns |     - |  283020 B |           - |
|               ToFrozenDictionary_Optimized |     #87876 | 10000 |              1 |    263,961.31 ns |  0.07 |  360130 B |        0.12 |
|               ToFrozenDictionary_Optimized |     #87688 | 10000 |              1 |  4,023,881.49 ns |  1.00 | 2942142 B |        1.00 |
|                                            |            |       |                |                  |       |           |             |
|                TryGetValue_True_Dictionary |     #87688 | 10000 |              1 | 12,874,506.46 ns |     - |         - |          NA |
| TryGetValue_True_FrozenDictionaryOptimized |     #87876 | 10000 |              1 |     65,572.38 ns |  1.08 |         - |          NA |
| TryGetValue_True_FrozenDictionaryOptimized |     #87688 | 10000 |              1 |     60,486.95 ns |  1.00 |         - |          NA |
|                                            |            |       |                |                  |       |           |             |
|                               ToDictionary |     #87688 | 10000 |              5 |    100,764.40 ns |     - |  283090 B |           - |
|               ToFrozenDictionary_Optimized |     #87876 | 10000 |              5 |    168,735.65 ns |  0.15 |  200128 B |        0.23 |
|               ToFrozenDictionary_Optimized |     #87688 | 10000 |              5 |  1,097,493.94 ns |  1.00 |  871771 B |        1.00 |
|                                            |            |       |                |                  |       |           |             |
|                TryGetValue_True_Dictionary |     #87688 | 10000 |              5 |  2,732,495.46 ns |     - |         - |          NA |
| TryGetValue_True_FrozenDictionaryOptimized |     #87876 | 10000 |              5 |    113,734.85 ns |  1.06 |         - |          NA |
| TryGetValue_True_FrozenDictionaryOptimized |     #87688 | 10000 |              5 |    107,283.76 ns |  1.00 |         - |          NA |